% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8all_results.R
\name{extract_all}
\alias{extract_all}
\title{MotionWatch 8 - Extraction de toutes les données}
\usage{
extract_all(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Liste : bases de données et variables.
}
\description{
Cette fonction fournit une liste des résultats des analyses de la MotionWatch 8.
}
\examples{
#extract_all()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
