% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8axial_mode.R
\name{extract_axial_mode}
\alias{extract_axial_mode}
\title{MotionWatch 8 - Extraction du mode}
\usage{
extract_axial_mode(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Chaîne de caractères.
}
\description{
Cette fonction extrait le mode utilisé par la MotionWatch 8.
}
\examples{
#extract_axial_mode()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
