% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8device_SN.R
\name{extract_device_SN}
\alias{extract_device_SN}
\title{MotionWatch 8 - Extraction du SN}
\usage{
extract_device_SN(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Chaîne de caractères.
}
\description{
Cette fonction extrait le numéro de série (SN) de la MotionWatch 8.
}
\examples{
#extract_device_SN()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
