% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8device_version.R
\name{extract_device_version}
\alias{extract_device_version}
\title{MotionWatch 8 - Extraction de la version}
\usage{
extract_device_version(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Chaîne de caractères.
}
\description{
Cette fonction extrait la version de la MotionWatch 8.
}
\examples{
#extract_device_version()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
