% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8epoch.R
\name{extract_epoch}
\alias{extract_epoch}
\title{MotionWatch 8 - Extraction de l'epoch}
\usage{
extract_epoch(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Nombre numérique.
}
\description{
Cette fonction extrait l'epoch définit lors de la programmation de la MotionWatch 8.
}
\examples{
#extract_epoch()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
