% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8frequency_light.R
\name{extract_frequency_light}
\alias{extract_frequency_light}
\title{MotionWatch 8 - Extraction de l'échantillonnage de la luminosité}
\usage{
extract_frequency_light(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Nombre numérique.
}
\description{
Cette fonction extrait la fréquence d'enregistrement des données de luminosité de la MotionWatch 8.
}
\examples{
#extract_frequency_light()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
