% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8light.R
\name{extract_light}
\alias{extract_light}
\title{MotionWatch 8 - Données de luminosité}
\usage{
extract_light(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Tableau de données
}
\description{
Cette fonction extrait les données de luminosité de la MotionWatch 8.
}
\examples{
#extract_light()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
