% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8motion.R
\name{extract_motion}
\alias{extract_motion}
\title{MotionWatch 8 - Données de mouvement (counts)}
\usage{
extract_motion(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Tableau de données
}
\description{
Cette fonction extrait les données de mouvement de la MotionWatch 8.
}
\examples{
#extract_motion()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
