% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8nap.R
\name{extract_nap}
\alias{extract_nap}
\title{MotionWatch 8 - Analyse des données de sieste}
\usage{
extract_nap(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Tableau de données
}
\description{
Cette fonction extrait les données de sieste de la MotionWatch 8.
}
\examples{
#extract_nap()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
