% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8npcra.R
\name{extract_npcra}
\alias{extract_npcra}
\title{MotionWatch 8 - Extraction des données npcra}
\usage{
extract_npcra(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Tableau de données.
}
\description{
Cette fonction extrait les résultats de l'analyse des rythmes circadiens non-paramétriques (npcra) de la MotionWatch 8.
}
\examples{
#extract_npcra()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
