% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8position.R
\name{extract_position}
\alias{extract_position}
\title{MotionWatch 8 - Extraction de la position}
\usage{
extract_position(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Chaîne de caractères.
}
\description{
Cette fonction extrait la position (poignet, cheville, gauche, droite...) définit lors de la programmation de la MotionWatch 8.
}
\examples{
#extractMW8position()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
