% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8sex.R
\name{extract_sex}
\alias{extract_sex}
\title{MotionWatch 8 - Extraction du sexe}
\usage{
extract_sex(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Chaîne de caractères.
}
\description{
Cette fonction extrait le sexe définit lors de la programmation de la MotionWatch 8.
}
\examples{
#extract_sex()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
