% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8sleep_analysis.R
\name{extract_sleep_analysis}
\alias{extract_sleep_analysis}
\title{MotionWatch 8 - Analyse du sommeil nuit par nuit}
\usage{
extract_sleep_analysis(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Tableau de données
}
\description{
Cette fonction extrait les résultats de l'analyse du sommeil nuit par nuit de la MotionWatch 8.
}
\examples{
#extract_sleep_analysis()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
