% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8start_recording.R
\name{extract_start_recording}
\alias{extract_start_recording}
\title{MotionWatch 8 - Extraction du début de l'enregistrement}
\usage{
extract_start_recording(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Date-heure.
}
\description{
Cette fonction extrait la date et l'heure de début de l'enregistrement par la MotionWatch 8.
}
\examples{
#extract_start_recording()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
