% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8TimeZone_Adjust.R
\name{extract_TZadj}
\alias{extract_TZadj}
\title{MotionWatch 8 - Extraction de l'ajustement du fuseau horaire}
\usage{
extract_TZadj(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Temps en heure(s).
}
\description{
Cette fonction extrait l'ajustement du fuseau horaire en heure(s) (à modifier dans MotionWare).
}
\examples{
#extract_TZadj()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
