% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8userID.R
\name{extract_userID}
\alias{extract_userID}
\title{MotionWatch 8 - Extraction de l'id}
\usage{
extract_userID(file)
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Chaîne de caractères.
}
\description{
Cette fonction extrait l'id définit lors de la programmation de la MotionWatch 8.
}
\examples{
#extract_userID()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
