% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8TimeZone_Adjust.R
\docType{data}
\name{file}
\alias{file}
\title{MotionWatch 8 - Extraction de l'ajustement du fuseau horaire}
\format{
An object of class \code{character} of length 1.
}
\usage{
file
}
\arguments{
\item{file}{Chemin du fichier .mtn}
}
\value{
Temps en heure(s).
}
\description{
Cette fonction extrait l'ajustement du fuseau horaire en heure(s) (à modifier dans MotionWare).
}
\examples{
#extract_TZadj()

}
\author{
European Sleep Center (Paris, France)

Maxime Chauvineau - mchauvineau@europeansleepcenter.fr
}
\keyword{datasets}
