% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW8process-package.R
\docType{package}
\name{MW8process-package}
\alias{MW8process}
\alias{MW8process-package}
\title{MW8process: MotionWatch8 data processing}
\description{
Process actigraphy data (MotionWatch8 device) from MotionWare software.
}
\author{
\strong{Maintainer}: Maxime Chauvineau \email{mchauvineau@europeansleepcenter.fr}

}
\keyword{internal}
